requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1

python-jose==3.3.0

supabase==2.10.0

google-genai==0.6.0

reportlab==4.2.5